requirements.txt
# Core dependencies
web3>=6.0.0
click>=8.0.0
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.8.0
eth-abi>=4.0.0

# AI dependencies
alith>=0.1.0
openai>=1.0.0

# Database
# Using built-in sqlite3

# Async support
# asyncio is built-in

# Optional dependencies for signature exchange
redis>=4.5.0  # For Redis-based signature exchange

# WebSocket support
websockets>=15.0.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0